+++ CHANGES.md
@@ -4,6 +4,8 @@
 Changes in 12.9.2
 -----------------
 - Fix SQL syntax error "AND STUDENT_ID IN()" when nothing to save in InputFinalGrades.php & EnterEligibility.php
+- SQL fix "Enroll student for next school year" when student is inactive in Enrollment.fnc.php
+
 Changes in 12.9.1
 -----------------
 - Fix double escaping Course Period Short Name in Courses.php & Courses.fnc.php